Added entrance and exit to room gen
Added function which places entrance and exit tiles in centre of walls depending on defined entrance and exit directions. The function also checks if  the door is blocked by a wall tile, if it is a new generation is made. The entrance and exit directions are defined when calling the StartGenerating function, this will allow me to pass the required entrance and exit positions based on the position in the overall level (maze). This is important to ensure the puzzle's entrance and/or exit is in the correct placement for the maze layout