Added A* algorithm
Added an A* algorithm to find a path between 2 points in the level grid (also discovers whether a path is impossible). This will be useful to check whether the player is able to navigate between the doors when the boxes are placed on the buttons. 
This still needs to be implemented.