Player centre issues fixed
Entrance and exit positions for first and last rooms are now set to the centre of the room. The generator now ensures that the centre tile of the first and last room are not wall tiles when generating the room. Also, the player is placed at the beginning of the puzzle generation to ensure boxes/buttons are not placed on top.